pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["xazme <juriii155@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.7.1"
fastapi = "^0.115.6"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
alembic = "^1.14.1"
bs4 = "^0.0.2"
requests = "^2.32.3"
pyjwt = {extras = ["cryptography"], version = "^2.10.1"}
cryptography = "^44.0.0"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "19.10b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
